+++ .samples/styles/styles.go
@@ -1,13 +1,12 @@
 package main
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/renatopp/go-cli"
-	"github.com/renatopp/go-cli/pkg"
-	cerrors "github.com/renatopp/go-cli/pkg/errors"
-	"github.com/renatopp/go-cli/pkg/formatters"
+	"github.com/renatopp/go-cli/core"
+	"github.com/renatopp/go-cli/errors"
+	"github.com/renatopp/go-cli/formatters"
 )
 
 func main() {
@@ -17,15 +16,15 @@ func main() {
 	cli.AutoHelp(true)
 
 	// Wrap the default help style with a custom banner.
-	cli.HelpFormatter(func(cmd *pkg.Command, loc pkg.Locale) string {
+	cli.HelpFormatter(func(cmd *core.Command, loc core.Locale) string {
 		return "== STYLES ==\n\n" + formatters.DefaultHelpFormatter(cmd, loc)
 	})
 
 	// Customize error messages, inspecting typed errors for special cases.
-	cli.ErrorFormatter(func(err error, loc pkg.Locale) string {
-		var cliErr *cerrors.CliError
+	cli.ErrorFormatter(func(err error, loc core.Locale) string {
+		var cliErr *errors.CliError
 		if errors.As(err, &cliErr) {
-			if cliErr.Code == cerrors.ErrUnknownFlag {
+			if cliErr.Code == errors.ErrUnknownFlag {
 				// Parameters: [0] = flag name
 				return fmt.Sprintf("oops! I don't know the flag %q", cliErr.Parameters[0])
 			}